requirements.txt
openai-whisper==20231117
torch>=1.9.0
torchaudio>=0.9.0
numpy>=1.21.0
librosa>=0.9.0
soundfile>=0.10.0
flask>=2.0.0
werkzeug>=2.0.0
